MicroHack-3.0-Back/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.20

# Scraping & Agent dependencies
langgraph==0.2.0
langchain==0.2.0
langchain-core==0.2.0
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.1.0
html5lib==1.1
httpx==0.26.0
pyyaml==6.0.1
minio==7.2.0